Extract volume override logic into StreamVolumeOverride
